backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching & Sessions
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.4
kombu==5.3.4

# AI & Integrations
anthropic==0.39.0
openai==1.3.8
google-generativeai==0.3.2
python-dateutil==2.8.2

# OAuth & External APIs
authlib==1.2.1
httpx==0.25.2

# Real-time Communication
python-socketio==5.10.0

# Templates & Static Files
jinja2==3.1.2
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.17.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Environment Management
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0